feat(test) : #25 테스트 코드 추가 구현 및 기능별 분리
